+++ CHANGELOG.md
@@ -2,13 +2,25 @@
 
 ## 2025-08-09
 
+**JS execution:**
+
 - `$foo("a", "b")` and other messages starting with `$` are now treated as JavaScript code to execute.
-- ...except for stuff like `$foo a b c` which is treated as a command invocation.
-- `$foo` can now be used in JS instead of `$.cmd("foo")`.
+- Also, `$ <arbitrary code>` is treated as JavaScript code to execute.
+- All commands are now available in JS as `$foo` objects.
+  - For nested commands, `foo/bar` is available as `$foo.bar`.
+    The old `$.cmd("foo/bar")` is still supported but deprecated.
+  - For commands with dashes, `foo-bar` is available as `$foo_bar`.
+
+**`$` prefix:**
+
+- When a message looks like a command and not as JS, e.g. `$hello` or `$rand 1 2`, it's treated as a command invocation.
+
+**Listing children:**
+
+- `$foo/`, `$foo.`, `$foo..` now list children of `$foo`.
 
 ## 2025-08-02
 
-- Added support for `$` prefix as an alternative to `.` for accessing aliases (e.g., `$weather` works the same as `.weather`).
 - BREAKING: Commands can access info about their own invocation via `this` instead of `context`.
 - BREAKING: `.show` shows command structure instead of running it.
 - Commands now have access to `$.message`, `$.user`, and `$.server`.